+++ backend/techpick-core/src/main/java/techpick/core/rabbitmq/EventSender.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import techpick.core.config.RabbitmqConfig;
 
 @Slf4j
 @Component
